requirements.txt
# coverage==7.5.0
django==4.2
django-cors-headers==3.13.0
django-db-logger==0.1.12
django-environ==0.9.0
django-filter==23.4
django-zxcvbn-password-validator==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.1
drf_standardized_errors==0.12.6
# factory-boy==3.3.0
gunicorn==20.1.0
# model-bakery-1.18.0
openai==0.27.2
psycopg2-binary==2.9.5